+++ src/components/TypingParagraphInputBox.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 
 interface Props {
-  inputRef: React.RefObject<HTMLInputElement>;
+  inputRef: React.RefObject<HTMLInputElement | null>;
   inputHandlers: {
     onInput: React.FormEventHandler<HTMLInputElement>;
     onKeyDown: React.KeyboardEventHandler<HTMLInputElement>;